- dfsound: quick fix (FF9 music, maybe others as well). - peopsxgl (unix): imported new config utility written by Kivutar, with minor amendments (https://github.com/Kivutar/cfgPeopsMesaGL).
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@61694 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/dfsound/externals.h
@@ -276,7 +276,7 @@ extern unsigned long  spuAddr;
 extern int      bEndThread;
 extern int      bThreadEnded;
 extern int      bSpuInit;
-extern unsigned long dwNewChannel;
+extern uint32_t dwNewChannel;
 
 extern int      SSumR[];
 extern int      SSumL[];
